+++ Resource/Source/Game/UI/MenuTopInf.cpp
@@ -22,8 +22,8 @@ MenuTopInf::MenuTopInf(const MapCtrl& mapCtrl, const unsigned char& turnCnt, std
 	GetGraphSize(turnCntFrameH, turnCntSize);
 
 	_moveAnimTrack = make_unique<Track<Vector2Int>>();
-	_moveAnimTrack->AddKey(0, Vector2Int(-nameFrameSize.w - turnCntSize.w, nameFrameSize.h * 0.5f + spaceY));
-	_moveAnimTrack->AddKey(30, Vector2Int(nameFrameSize.w * 0.5f, nameFrameSize.h * 0.5f + spaceY));
+	_moveAnimTrack->AddKey(0, Vector2Int(-nameFrameSize.w - turnCntSize.w, static_cast<int>(nameFrameSize.h * 0.5f) + spaceY));
+	_moveAnimTrack->AddKey(30, Vector2Int(static_cast<int>(nameFrameSize.w * 0.5f), static_cast<int>(nameFrameSize.h * 0.5f) + spaceY));
 	_moveAnimTrack->SetReverse(true);
 	_moveAnimTrack->End();
 }
@@ -51,7 +51,7 @@ void MenuTopInf::Draw()
 	GetGraphSize(turnCntFrameH, turnCntSize);
 
 	Rect nameFrameRect(_moveAnimTrack->GetValue(), nameFrameSize);
-	Rect turnCntFrameRect(Vector2Int(nameFrameRect.Right() + turnCntSize.w * 0.5f, nameFrameRect.center.y), turnCntSize);
+	Rect turnCntFrameRect(Vector2Int(nameFrameRect.Right() + static_cast<int>(turnCntSize.w * 0.5f), nameFrameRect.center.y), turnCntSize);
 
 	nameFrameRect.DrawGraph(nameFrameH);
 	turnCntFrameRect.DrawGraph(turnCntFrameH);
